# docs/swagger/swagger.yaml
basePath: /
definitions:
...
  helpers.Response:
    properties:
      data: {}
      error: {}
      status:
        type: integer
      success:
        type: boolean
    type: object
...
  ws_auth.refreshRequest:
    properties:
      refresh_token:
        type: string
    required:
    - refresh_token
    type: object
info:
  contact: {}
  description: Go API service
  title: Go API
  version: "1.0"
paths:
...
  /storages/ws-token:
    post:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/helpers.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/helpers.Response'
        "429":
          description: Too Many Requests
          schema:
            $ref: '#/definitions/helpers.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/helpers.Response'
      security:
      - BearerAuth: []
      summary: Issue websocket token pair
      tags:
      - ws-auth
  /storages/ws-token/refresh:
    post:
      consumes:
      - application/json
      parameters:
      - description: Request body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/ws_auth.refreshRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/helpers.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helpers.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/helpers.Response'
        "429":
          description: Too Many Requests
          schema:
            $ref: '#/definitions/helpers.Response'
      summary: Rotate websocket token pair
      tags:
      - ws-auth
...
  /whatsapp/login:
    post:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/helpers.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/helpers.Response'
      security:
      - BearerAuth: []
      summary: Trigger WhatsApp login QR generation
      tags:
      - whatsapp
  /whatsapp/logout:
    post:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/helpers.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/helpers.Response'
      security:
      - BearerAuth: []
      summary: Logout WhatsApp
      tags:
      - whatsapp
  /whatsapp/qr-code:
    get:
      produces:
      - image/png
      responses:
        "200":
          description: OK
          schema:
            type: file
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/helpers.Response'
      security:
      - BearerAuth: []
      summary: Get WhatsApp QR code
      tags:
      - whatsapp
  /whatsapp/status:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/helpers.Response'
      security:
      - BearerAuth: []
      summary: Get WhatsApp status
      tags:
      - whatsapp
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by the access token
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"